Change all occurences of 'video' to ' lesson' in LessonList storybook file
